// front-end-app/quanly-thuysan/src/app/model/nuoigiong.ts
export class NuoiGiong {
  $class: string;
  idgdng: string;
  sanphamnoi: string;
  sanphamngoai: string;
  nguyenlieu: string;
  ao: string;
  trangtrai: string;
  tgnuoi: string;
  tgxuatao: string;
  phuongtien: string;
  trangtrainhan: string;
  soluong: number;
  mota: string;
  trangtrainuoigiong: string;
}